pyproject.toml
[tool.poetry]
name = "clsp"
version = "0.1.0"
description = ""
authors = ["Niels Horn <niels.horn@unity3d.com>"]
readme = "README.md"
packages = [
    { include = "clsp" },
]

[tool.poetry.dependencies]
python = "~3.10"
black = "^23.1.0"
torch = "^1.13.1"
isort = "^5.12.0"
einops = "^0.6.0"
numpy = "^1.24.2"
pytest = "^7.2.1"
google-cloud-storage = "^2.7.0"
torchaudio = "^0.13.1"
google-cloud-bigquery = "^3.5.0"
polars = "^0.16.6"
tqdm = "^4.64.1"
transformers = "^4.26.1"
tiktoken = "^0.2.0"
openai-whisper = "^20230124"


[tool.poetry.group.dev.dependencies]
ipython = "^8.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
